refactor: Remove ContentModerationService and relocate remaining domain services
- Removed unused `ContentModerationService` due to lack of implementation.
- Relocated `TaggingService` from `Application.DomainServices` to `Application.Services` for consistency.
- Updated namespaces and references across the project to reflect new service locations.
- Removed unused `MediaAttachmentService` registration and relevant references.
