+++ server/ExpertBridge.Admin/Components/Pages/ModerationReports/ModerationReportDetail.razor.cs
@@ -291,8 +291,8 @@ public partial class ModerationReportDetail : ComponentBase
         }
 
         var confirmationMessage = $"Are you sure you want to mark this {_report.ContentType} as {(_report.IsNegative ?
-                                      $"Clean and undelete the {_report.ContentType}" :
-                                      $"Negative and delete the {_report.ContentType}")}?";
+            $"Clean and undelete the {_report.ContentType}" :
+            $"Negative and delete the {_report.ContentType}")}?";
         var confirmed = await DialogService.Confirm(
             confirmationMessage,
             "Confirm",
@@ -349,9 +349,9 @@ public partial class ModerationReportDetail : ComponentBase
                 await Cache.RemoveAsync($"admin:moderation-reports:{Id}");
 
                 _report.IsNegative = !_report.IsNegative;
-                var message = _report.IsNegative ?
-                    $"{_report.ContentType} marked as negative and deleted" :
-                    $"{_report.ContentType} marked as clean and restored";
+                var message = _report.IsNegative
+                    ? $"{_report.ContentType} marked as negative and deleted"
+                    : $"{_report.ContentType} marked as clean and restored";
                 ShowNotification(NotificationSeverity.Success, "Success", message);
             }
             catch (Exception ex)
@@ -441,10 +441,7 @@ public partial class ModerationReportDetail : ComponentBase
     {
         NotificationService.Notify(new NotificationMessage
         {
-            Severity = severity,
-            Summary = summary,
-            Detail = detail,
-            Duration = 4000
+            Severity = severity, Summary = summary, Detail = detail, Duration = 4000
         });
     }
 }
