+++ server/ExpertBridge.Api/Controllers/MessagesController.cs
@@ -1,5 +1,4 @@
 ﻿using ExpertBridge.Api.Services;
-using ExpertBridge.Application.DomainServices;
 using ExpertBridge.Contract.Requests.CreateMessage;
 using ExpertBridge.Contract.Responses;
 using Microsoft.AspNetCore.Authorization;
