+++ server/ExpertBridge.Api/Controllers/PostsController.cs
@@ -1,5 +1,4 @@
 using ExpertBridge.Api.Services;
-using ExpertBridge.Application.DomainServices;
 using ExpertBridge.Application.Settings;
 using ExpertBridge.Contract.Requests.CreatePost;
 using ExpertBridge.Contract.Requests.EditPost;
