+++ server/ExpertBridge.Api/Controllers/ProfilesController.cs
@@ -1,5 +1,4 @@
 using ExpertBridge.Api.Services;
-using ExpertBridge.Application.DomainServices;
 using ExpertBridge.Application.Helpers;
 using ExpertBridge.Application.Settings;
 using ExpertBridge.Contract.Requests.OnboardUser;
