+++ server/ExpertBridge.Api/Services/CommentService.cs
@@ -1,4 +1,4 @@
-﻿using ExpertBridge.Application.DomainServices;
+﻿using ExpertBridge.Application.Services;
 using ExpertBridge.Contract.Messages;
 using ExpertBridge.Contract.Queries;
 using ExpertBridge.Contract.Requests.CreateComment;
@@ -201,7 +201,7 @@ public class CommentService
     ///     }
     /// };
     /// var comment = await commentService.CreateCommentAsync(request, currentProfile);
-    ///
+    /// 
     /// // Nested reply to existing comment
     /// var replyRequest = new CreateCommentRequest {
     ///     PostId = "post123",
@@ -402,7 +402,7 @@ public class CommentService
     ///     <code>
     /// // Anonymous user viewing comment
     /// var comment = await commentService.GetCommentAsync("comment123", null);
-    ///
+    /// 
     /// // Authenticated user viewing comment (includes their vote status)
     /// var comment = await commentService.GetCommentAsync("comment123", currentUser.ProfileId);
     /// if (comment.CurrentUserVote == VoteType.Upvote) {
@@ -447,7 +447,7 @@ public class CommentService
     ///     **Example Usage:**
     ///     <code>
     /// var comments = await commentService.GetCommentsByPostAsync("post123", currentUser.ProfileId);
-    ///
+    /// 
     /// // Build comment tree client-side
     /// var topLevel = comments.Where(c => c.ParentCommentId == null);
     /// foreach (var comment in topLevel) {
@@ -559,7 +559,7 @@ public class CommentService
     ///     profileId,
     ///     currentUser.ProfileId
     /// );
-    ///
+    /// 
     /// // Show statistics
     /// var totalComments = userComments.Count;
     /// var totalUpvotes = userComments.Sum(c => c.UpvotesCount);
@@ -642,7 +642,7 @@ public class CommentService
     /// var editRequest = new EditCommentRequest {
     ///     Content = "Updated comment with additional context."
     /// };
-    ///
+    /// 
     /// try {
     ///     var updated = await commentService.EditCommentAsync(
     ///         "comment123",
@@ -753,7 +753,7 @@ public class CommentService
     ///     **Vote Processing Logic:**
     ///     <code>
     /// existingVote = Get current user's vote on this comment
-    ///
+    /// 
     /// if (existingVote == null) {
     ///     // No vote yet, create new
     ///     CreateCommentVote(commentId, voterId, isUpvoteIntent)
@@ -781,7 +781,7 @@ public class CommentService
     /// );
     /// // updated.UpvotesCount incremented
     /// // updated.CurrentUserVote == VoteType.Upvote
-    ///
+    /// 
     /// // User clicks upvote again (toggle off)
     /// updated = await commentService.VoteCommentAsync(
     ///     "comment123",
@@ -912,7 +912,7 @@ public class CommentService
     /// Comment B (reply to A)
     ///     ↓ ParentCommentId
     /// Comment C (reply to B)
-    ///
+    /// 
     /// DELETE Comment A → Cascade deletes B and C
     /// </code>
     ///     **Cascade Configuration:**
