--- server/ExpertBridge.Application/DomainServices/ContentModerationService.cs
@@ -1,210 +1,0 @@
-﻿using ExpertBridge.Contract.Responses;
-using ExpertBridge.Data.DatabaseContexts;
-
-namespace ExpertBridge.Application.DomainServices;
-
-/// <summary>
-///     Provides content moderation services for detecting and handling inappropriate content in posts and comments.
-/// </summary>
-/// <remarks>
-///     This service is currently a placeholder/stub for future implementation of content moderation features.
-///     **Planned Features:**
-///     - Automated inappropriate language detection reporting
-///     - Manual user-submitted content reports
-///     - Integration with AI-powered content moderation (Groq LLM)
-///     - Email notifications to administrators for flagged content
-///     - Content flagging and review workflow
-///     - User reputation impact from moderation actions
-///     **Architecture Integration:**
-///     Content moderation works in conjunction with:
-///     - Background workers that analyze content asynchronously
-///     - Message broker (MassTransit/RabbitMQ) for queuing moderation tasks
-///     - AI services (AiPostTaggingService) for detecting inappropriate language
-///     - Notification system for alerting moderators and users
-///     **Future Implementation Approach:**
-///     **1. Automated Detection Flow:**
-///     - Post/Comment created → Published to DetectInappropriateContentMessage
-///     - Background consumer calls Groq API for toxicity analysis
-///     - If flagged: Call ReportPostAsync/ReportCommentAsync with AI results
-///     - Store moderation record in database
-///     - Notify moderators via email/push notification
-///     **2. Manual Reporting Flow:**
-///     - User reports content via "Report" button
-///     - Call ReportPostAsync/ReportCommentAsync with user reason
-///     - Create moderation ticket for review
-///     - Notify moderators
-///     **3. Moderation Actions:**
-///     - Soft delete flagged content (ISoftDeletable pattern)
-///     - Suspend user accounts temporarily
-///     - Reduce user reputation points
-///     - Send warning notifications to content authors
-///     **Database Schema (Planned):**
-///     <code>
-/// public class ContentModerationReport
-/// {
-///     public string Id { get; set; }
-///     public string PostId { get; set; }
-///     public string CommentId { get; set; }
-///     public string ReportedBy { get; set; } // User ID or "System" for AI
-///     public string Reason { get; set; }
-///     public double ToxicityScore { get; set; }
-///     public ModerationStatus Status { get; set; } // Pending, Reviewed, Actioned
-///     public DateTime ReportedAt { get; set; }
-///     public string ReviewedBy { get; set; }
-///     public string ActionTaken { get; set; }
-/// }
-/// </code>
-///     **Example Integration:**
-///     <code>
-/// // In background consumer after AI detection
-/// if (toxicityScore > 0.7)
-/// {
-///     await moderationService.ReportPostAsync(
-///         postId,
-///         aiDetectionResults,
-///         isNegative: true
-///     );
-/// }
-/// 
-/// // User-initiated report
-/// await moderationService.ReportCommentAsync(
-///     commentId,
-///     "This comment contains hate speech"
-/// );
-/// </code>
-///     **Configuration (Future):**
-///     <code>
-/// "ContentModeration": {
-///   "ToxicityThreshold": 0.7,
-///   "AutoDeleteThreshold": 0.9,
-///   "ModeratorEmails": ["admin@expertbridge.com"],
-///   "EnableAutoModeration": true
-/// }
-/// </code>
-///     The service is registered as scoped in the DI container and will be fully implemented
-///     when content moderation requirements are finalized.
-/// </remarks>
-public class ContentModerationService
-{
-    private readonly ExpertBridgeDbContext _dbContext;
-
-    /// <summary>
-    ///     Initializes a new instance of the <see cref="ContentModerationService" /> class.
-    /// </summary>
-    /// <param name="dbContext">The database context for accessing content and moderation data.</param>
-    public ContentModerationService(ExpertBridgeDbContext dbContext)
-    {
-        // Get a channel to an email sending service
-        // to delegate email sending to it.
-        _dbContext = dbContext;
-    }
-
-    /// <summary>
-    ///     Reports a post based on automated inappropriate language detection results.
-    /// </summary>
-    /// <param name="postId">The unique identifier of the post to report.</param>
-    /// <param name="results">
-    ///     The AI-powered language detection results containing toxicity scores and flagged content categories.
-    /// </param>
-    /// <param name="isNegative">
-    ///     Indicates whether the content is determined to be negative/inappropriate based on analysis.
-    /// </param>
-    /// <returns>A task representing the asynchronous reporting operation.</returns>
-    /// <remarks>
-    ///     **Planned Implementation:**
-    ///     - Store moderation report in database with AI detection results
-    ///     - Create notification for moderators if toxicity exceeds threshold
-    ///     - Optionally auto-hide content if confidence score is very high
-    ///     - Update post metadata to mark as flagged/under review
-    ///     - Send email to administrators for review
-    ///     **Use Case:**
-    ///     Called by background workers after AI content analysis detects potential violations.
-    ///     **Future Enhancement:**
-    ///     May integrate with reputation system to automatically penalize repeat offenders.
-    /// </remarks>
-    public async Task ReportPostAsync(
-        string postId,
-        InappropriateLanguageDetectionResponse results,
-        bool isNegative)
-    {
-    }
-
-    /// <summary>
-    ///     Reports a post with a user-provided reason for manual moderator review.
-    /// </summary>
-    /// <param name="postId">The unique identifier of the post to report.</param>
-    /// <param name="reason">
-    ///     The user-provided explanation for why the post is being reported
-    ///     (e.g., "Spam", "Harassment", "Misinformation").
-    /// </param>
-    /// <returns>A task representing the asynchronous reporting operation.</returns>
-    /// <remarks>
-    ///     **Planned Implementation:**
-    ///     - Validate postId exists and is not already deleted
-    ///     - Create moderation ticket with user-provided reason
-    ///     - Add to moderation queue for admin review
-    ///     - Send confirmation notification to reporting user
-    ///     - Notify moderators via email/dashboard
-    ///     - Track reporting user to prevent abuse of reporting system
-    ///     **Use Case:**
-    ///     Called when users click "Report" button on inappropriate posts.
-    ///     **Anti-Abuse:**
-    ///     Should implement rate limiting to prevent spam reporting.
-    /// </remarks>
-    public async Task ReportPostAsync(string postId, string reason)
-    {
-    }
-
-    /// <summary>
-    ///     Reports a comment based on automated inappropriate language detection results.
-    /// </summary>
-    /// <param name="commentId">The unique identifier of the comment to report.</param>
-    /// <param name="results">
-    ///     The AI-powered language detection results containing toxicity scores and flagged content categories.
-    /// </param>
-    /// <returns>A task representing the asynchronous reporting operation.</returns>
-    /// <remarks>
-    ///     **Planned Implementation:**
-    ///     - Store moderation report with AI analysis results
-    ///     - Flag comment for moderator review if toxicity exceeds threshold
-    ///     - Optionally auto-delete if confidence is extremely high (>0.95)
-    ///     - Update comment metadata to mark as under review
-    ///     - Notify comment author of flagged content
-    ///     - Send email alert to moderators
-    ///     **Use Case:**
-    ///     Called by background workers after DetectInappropriateCommentMessage is processed.
-    ///     **Integration Point:**
-    ///     Works with CommentService which publishes DetectInappropriateCommentMessage after comment creation/edit.
-    /// </remarks>
-    public async Task ReportCommentAsync(
-        string commentId,
-        InappropriateLanguageDetectionResponse results)
-    {
-    }
-
-    /// <summary>
-    ///     Reports a comment with a user-provided reason for manual moderator review.
-    /// </summary>
-    /// <param name="commentId">The unique identifier of the comment to report.</param>
-    /// <param name="reason">
-    ///     The user-provided explanation for why the comment is being reported
-    ///     (e.g., "Offensive language", "Personal attack", "Off-topic").
-    /// </param>
-    /// <returns>A task representing the asynchronous reporting operation.</returns>
-    /// <remarks>
-    ///     **Planned Implementation:**
-    ///     - Validate commentId exists and is not already deleted
-    ///     - Create moderation report entry in database
-    ///     - Add to moderation review queue
-    ///     - Send notification to reporting user confirming receipt
-    ///     - Alert moderators via email and dashboard notification
-    ///     - Track reporting patterns to identify malicious reporters
-    ///     **Use Case:**
-    ///     Called when users report inappropriate comments manually.
-    ///     **Future Enhancement:**
-    ///     Aggregate multiple reports on same comment to prioritize moderator review.
-    /// </remarks>
-    public async Task ReportCommentAsync(string commentId, string reason)
-    {
-    }
-}
