+++ server/ExpertBridge.Application/Services/AiPostTaggingService.cs
@@ -117,7 +117,7 @@ public sealed class AiPostTaggingService
     ///             "I'm looking for best practices to improve React app performance...",
     ///             new[] { "react", "javascript" }
     ///         );
-    ///
+    /// 
     ///         // Result might include:
     ///         // Language: English
     ///         // Tags: ["react", "javascript", "performance optimization", "web development", "frontend"]
