+++ server/ExpertBridge.Worker/Consumers/PostTaggingConsumer.cs
@@ -1,7 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using ExpertBridge.Application.DomainServices;
 using ExpertBridge.Application.Services;
 using ExpertBridge.Contract.Messages;
 using MassTransit;
